src/project/requirements.txt
django<=4.2.4
autopep8<=2.0.2
pycodestyle<=2.10.0
psycopg2-binary<=2.9.7
# django-bootstrap-v5<=1.0.11
djangorestframework==3.14.0
wui-python-additions==0.1.2
django-import-export<=3.3.1
python-dotenv==1.0.0
djangorestframework-simplejwt==5.3.1